+++ src/test/java/com/google/sps/api/CalendarClientHelperTest.java
@@ -16,10 +16,10 @@ package com.google.sps.api;
 
 import com.google.api.services.calendar.model.Event;
 import com.google.api.services.calendar.model.EventAttendee;
+import com.google.sps.api.Calendar.CalendarClientHelper;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import java.io.IOException;
